Move unit tests onto new build
