+++ applications/dashboard/src/scripts/__tests__/setup.ts
@@ -6,9 +6,8 @@
 
 import Enzyme from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import { importAll } from "@testroot/utility";
+import { importAll } from "@library/__tests__/utility";
 
 // Setup enzyme
 Enzyme.configure({ adapter: new Adapter() });
-
 importAll((require as any).context("..", true, /.test.(ts|tsx)$/));
