+++ plugins/rich-editor/src/scripts/__tests__/setup.ts
@@ -6,7 +6,7 @@
 
 import Enzyme from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import { importAll } from "@testroot/utility";
+import { importAll } from "@library/__tests__/utility";
 import reducerRegistry from "@library/state/reducerRegistry";
 import editorReducer from "@rich-editor/state/editorReducer";
 import registerQuill from "@rich-editor/quill/registerQuill";
